Update `make_better_change` instructions
Previous instructions implied that all coins should be used in each successive recursive call, which generates all change permutations and is in direct conflict with the Nota Bene ('Don't generate every possible permutation of coins and then compare them'). This is definitely confusing for students that followed the instructions, but cannot understand why their code is so slow.